+++ gladier/tests/test_state_models.py
@@ -4,15 +4,14 @@ import typing as t
 
 import pytest
 from gladier import (
-    ActionState,
-    ActionExceptionName,
     GladierBaseClient,
-    GladierClient,
     GladierBaseState,
     GladierBaseTool,
+    GladierClient,
     StateWithNextOrEnd,
     generate_flow_definition,
 )
+from gladier.tools.builtins import ActionExceptionName, ActionState
 
 
 def mock_func(data):
